requirements.txt
app==0.0.1
bokeh==2.4.2
cleandata==0.0.0
engines==0.0.1
features==0.5.12
filters==1.3.2
matplotlib==3.5.1
numpy==1.21.5
openpyxl==3.0.9
pandas==1.4.2
Pillow==9.2.0
plotly_express==0.4.1
statsmodels==0.13.2
streamlit==1.12.2
streamlit_bokeh_events==0.1.2
streamlit_option_menu==0.3.2
streamlit_plotly_events==0.0.6
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
